Writing to csv files number of data points
I initially wanted the map to dynamically update with a table of internet connections per region/ urban rural area. I couldn't get this working, it is something I may return to.

In the end I decided to just write to csv files the information that would be displayed. This was for a number of reasons, including the reason above. Basically it would add more loading time to perform the over() function every time the user changed an input, and there were only 12 potential combinations (6 datapoint files, 2 shapefiles) to display. If there were more, or the user were to be able to upload additional files I think I would have to look into the application being able to dynamically update itself.

For now, there are other parts of the project I need to move on to, so this means the map is 99% finished. Beyond the project I may try to create the desired functionality.